Add mutation support into CompositedCompiler and reorganize for clarity
